Kill styled components in favor of CSS Modules
More often than not, everything can be more manageable with
carefully written plain old CSS files. (Even better when enhanced
by CSS Modules)

Styled components offer a couple of great benefits but at the cost
of a performance overhead, security risks and a dangerous paradigm
shift. (Dangerous as in making it very easy to sprinkle ad hoc
styles everywhere and tightly couple them with the component logic
in the most unimaginable ways; along with implicitly encouraging
the creation of more components than actually necessary.)

Since we don't need those "couple of benefits" in this project, we
can go back to basics for the time being.